make loading independently
