+++ src/component/MusicLayout.js
@@ -1,40 +1,9 @@
 import React, { useState, useRef, useEffect } from 'react';
 import Navbar from './Navbar';
-import anime from 'animejs';
+import Loading from './shared/Loading';
 
 const MusicLayout = ({ component }) => {
-  const ref = useRef();
   const [loading, setLoading] = useState(true);
-  const [up, setUp] = useState(false);
-  const [down, setDown] = useState(false);
-
-  const moveUp = () => {
-    return anime({
-      targets: '.anime-character',
-      translateY: [-1000, 10, 10],
-      loop: true,
-      direction: 'alternate',
-      delay: function (el, i) {
-        return i * 100;
-      },
-      easing: 'easeInOutSine',
-    });
-  };
-
-  setTimeout(() => {
-    moveUp();
-  });
-
-  const wordSplitter = () =>
-    'LOADING...'.split('').map((char) => {
-      console.log(char);
-      return (
-        <span ref={ref} className="anime-character">
-          {char}
-        </span>
-      );
-    });
-
   return (
     <div className="embed-container">
       <iframe
@@ -42,9 +11,7 @@ const MusicLayout = ({ component }) => {
         onLoad={() => setLoading(false)}
       ></iframe>
       {loading ? (
-        <div ref={ref} className="char-container">
-          {wordSplitter()}
-        </div>
+        <Loading />
       ) : (
         <>
           <Navbar />
